added login, signup, home page, moved app page to home, added router functions
