+++ client/src/components/Nav.jsx
@@ -2,6 +2,17 @@ export default function Nav() {
     return (
         <nav>
           <h1 className="text-3xl text-center font-bold">Tweeter</h1>
+          <ul>
+            <li>
+              <a href="/">Home</a>
+            </li>
+            <li>
+              <a href="/login">Login</a>
+            </li>
+            <li>
+              <a href="/signup">Signup</a>
+            </li>
+          </ul>
         </nav>
     )
 }
