+++ client/src/main.jsx
@@ -1,11 +1,61 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
+// import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import App from './App.jsx'
 import './css/index.css'
 
+import Home from './pages/Home.jsx';
+import Login from './pages/Login.jsx';
+import Signup from './pages/Signup.jsx';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    children: [
+      {
+        index: true,
+        element: <Home />
+      }, 
+      {
+        path: 'login',
+        element: <Login />
+      },
+      {
+        path: 'signup',
+        element: <Signup />
+      }, 
+      {
+        path: 'signup',
+        element: <Signup />
+      }, 
+    ]
+  }
+]);
+
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-)
- 
+    <RouterProvider router={router} />
+  </React.StrictMode>
+);
+
+// ReactDOM.createRoot(document.getElementById('root')).render(
+//   <React.StrictMode>
+//     <Router>
+//       <Routes>
+//         <Route path="/" element={<App />}>
+//           <Route index element={<Home />} />
+//           <Route path="login" element={<Login />} />
+//           <Route path="signup" element={<Signup />} />
+//         </Route>
+//       </Routes>
+//     </Router>
+//   </React.StrictMode>,
+// );
+
+// ReactDOM.createRoot(document.getElementById('root')).render(
+//   <React.StrictMode>
+//     <App />
+//   </React.StrictMode>,
+// )
