+++ client/src/utils/mutations.js
@@ -1,0 +1,25 @@
+import { gql } from '@apollo/client';
+
+export const ADD_USER = gql`
+    mutation addUser($username: String!, $email: String!, $password: String!) {
+        addUser(username: $username, email: $email, password: $password) {
+        token
+        user {
+            _id
+            username
+        }
+        }
+    }
+`;
+
+export const LOGIN_USER = gql`
+    mutation login($email: String!, $password: String!) {
+        login(email: $email, password: $password) {
+        token
+        user {
+            _id
+            username
+        }
+        }
+    }
+`;
